Refactor: relocate intersection observer function to NewsList.js
