+++ src/App.js
@@ -1,6 +1,6 @@
 import Component from './components/core/Component.js';
 import { getNews } from './common/api/news.js';
-import { ALTERNATIVE } from './common/constants.js';
+import { ALTERNATIVE, CATEGORIES, POSITION } from './common/constants.js';
 import { $ } from './common/utils.js';
 import Nav from './components/Nav.js';
 import NavList from './components/NewsList.js';
@@ -12,9 +12,9 @@ export default class App extends Component {
 
   async initialState() {
     this.setState({
-      defaultCategory: await getNews('all'),
+      defaultCategory: await getNews(CATEGORIES.GENERAL.EN),
       activeCategory: null,
-      categoryName: 'all',
+      categoryName: CATEGORIES.GENERAL.EN,
     });
   }
 
@@ -35,37 +35,6 @@ export default class App extends Component {
       activeCategory: this.state?.activeCategory,
       categoryName: this.state?.categoryName,
     });
-
-    const observer = new IntersectionObserver((entries) => {
-      const page = Math.floor($('.news-list').children.length / 5) + 1;
-
-      // 최대 로딩 가능한 페이지 수  5 로 설정
-      let total = 5;
-
-      if (page === total) return;
-
-      entries.forEach(async (entry) => {
-        if (!entry.isIntersecting) return;
-
-        const articles = await getNews(this.state.categoryName, page);
-
-        articles.map((article) => {
-          let { url, urlToImage, title, description } = article;
-
-          if (!urlToImage) urlToImage = 'img/img-not-found.jpeg';
-
-          if (!description) description = ALTERNATIVE.DESCRIPTION;
-
-          this.loadMoreNews(
-            $('.news-list'),
-            'beforeend',
-            this.newsTemplate(url, urlToImage, title, description)
-          );
-        });
-      });
-    });
-
-    observer.observe($('.scroll-observer'));
   }
 
   async handleNav(target) {
