+++ src/components/core/Component.js
@@ -17,8 +17,6 @@ export default class Component {
   setState(newState) {
     this.state = newState;
     this.render();
-
-    console.log('state', this.state);
   }
 
   template() {
